# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Malaika-booking API",
        "description": "",
        "contact": {
            "email": "kirian@fruntera.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lista de clientes",
                "description": "Lista los clientes",
                "operationId": "clientList",
                "parameters": [
                    {
                        "name": "client_type",
                        "in": "query",
                        "description": "Id de tipo de cliente",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre del cliente",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "surname",
                        "in": "query",
                        "description": "Apellido del cliente",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Teléfono del cliente",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email del cliente",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dni",
                        "in": "query",
                        "description": "DNI del cliente",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "passport",
                        "in": "query",
                        "description": "Número de pasaporte del cliente",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client list retrieves successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ClientListResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Crea un nuevo cliente",
                "description": "Crea un nuevo cliente",
                "operationId": "createClient",
                "requestBody": {
                    "description": "Create user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "warehouse_id",
                                    "stock"
                                ],
                                "properties": {
                                    "client_type_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "language_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Selene"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Selenita"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "6945798"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "selene@gmail.com"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "47854123X"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Caella falsa 123, 08194, Barcelona"
                                    },
                                    "dni_expiration": {
                                        "type": "string",
                                        "example": "2025/12/01"
                                    },
                                    "place_birth": {
                                        "type": "string",
                                        "example": "Barcelona"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client created successfully"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Retorna un cliente por ID",
                "description": "Retorna un cliente",
                "operationId": "getClientById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client retrived successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ClientResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Actualiza los datos del cliente",
                "description": "Actualiza los datos del cliente",
                "operationId": "updateClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update client",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "warehouse_id",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Selene"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Selenita"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "6945798"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "selene@gmail.com"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "47854123X"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Caella falsa 123, 08194, Barcelona"
                                    },
                                    "dni_expiration": {
                                        "type": "string",
                                        "example": "2025/12/01"
                                    },
                                    "place_birth": {
                                        "type": "string",
                                        "example": "Barcelona"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client updated successfully"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClientListResource": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "surname": {
                        "description": "Client surname",
                        "type": "string",
                        "example": "Birgulilla"
                    },
                    "name": {
                        "description": "Client name",
                        "type": "string",
                        "example": "Maria"
                    },
                    "phone": {
                        "description": "Client phone",
                        "type": "string",
                        "example": "648595623"
                    },
                    "email": {
                        "description": "Client email",
                        "type": "string",
                        "example": "mariabirgulilla@gmail.com"
                    },
                    "DNI": {
                        "description": "Client DNI",
                        "type": "string",
                        "example": "47864512"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ClientListResource"
                }
            },
            "ClientResource": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "surname": {
                        "description": "Client surname",
                        "type": "string",
                        "example": "Birgulilla"
                    },
                    "name": {
                        "description": "Client name",
                        "type": "string",
                        "example": "Maria"
                    },
                    "phone": {
                        "description": "Client phone",
                        "type": "string",
                        "example": "648595623"
                    },
                    "email": {
                        "description": "Client email",
                        "type": "string",
                        "example": "mariabirgulilla@gmail.com"
                    },
                    "DNI": {
                        "description": "Client DNI",
                        "type": "string",
                        "example": "47864512"
                    },
                    "address": {
                        "description": "Client address",
                        "type": "string",
                        "example": "Calle falsa 123, 08167, Barcelona"
                    },
                    "dni_expiration": {
                        "description": "Client DNI expiration date",
                        "type": "string",
                        "example": "2025-12-01"
                    },
                    "place_birth": {
                        "description": "Client place of birtgh",
                        "type": "string",
                        "example": "Barcelona"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ClientListResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}